requirements.txt
flask
flask-login
werkzeug
numpy
fpdf2
gunicorn
python-dotenv
